:sparkles: Add organization service
